Migrate to TanStack Router with mobile-friendly compact layout
Replace state-based page switching with TanStack Router for proper URL
routing. Home page (/) is now the study page. Profile and settings are
accessible via a hamburger menu icon, with back buttons on sub-routes.
Reset functionality moved into settings page. Header significantly
compacted for mobile devices.
