+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@tanstack/react-query": "^5.95.2",
+    "@tanstack/react-router": "^1.168.10",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "wanakana": "^5.3.1",
