+++ src/main.tsx
@@ -1,7 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import App from './App'
+import { RouterProvider } from '@tanstack/react-router'
+import { router } from './router'
 import './index.css'
 
 if ('serviceWorker' in navigator) {
@@ -13,7 +14,7 @@ const queryClient = new QueryClient()
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
-      <App />
+      <RouterProvider router={router} />
     </QueryClientProvider>
   </StrictMode>
 )
