+++ src/pages/ProfileRoute.tsx
@@ -1,0 +1,8 @@
+import { ProfilePage } from '../components/ProfilePage'
+import { useVocabContext } from '../context/VocabularyContext'
+
+export function ProfileRoute() {
+  const { words, activeLang, activeLevel } = useVocabContext()
+
+  return <ProfilePage words={words} activeLang={activeLang} activeLevel={activeLevel} />
+}
